+++ shared/src/commonMain/kotlin/com/zhenxiang/bithelper/shared/model/WithdrawMethod.kt
@@ -9,7 +9,7 @@ data class WithdrawMethod(
     val hints: String?,
     val available: Boolean,
     val fee: BigDecimal?,
-    val decimalPrecision: Long,
+    val decimalPrecision: Int,
     val minAmount: BigDecimal,
     val hasMemoField: Boolean?,
 )
